Created separate RegistrationViewModel, fixed naming
