+++ Moodis/Moodis/Feature/Login/Register/RegisterViewModel.cs
@@ -2,18 +2,17 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Moodis.Feature.Login
+namespace Moodis.Feature.Login.Register
 {
-    class SignInViewModel
+    class RegisterViewModel
     {
         public static List<User> userList;
         public static User currentUser;
 
-        public User Authenticate(string username, string password)
+        public bool AddUser(string username, string password)
         {
             userList = new List<User>();
 
@@ -21,8 +20,17 @@ namespace Moodis.Feature.Login
             {
                 userList = Serializer.Load<List<User>>(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "/users.bin");
             }
-            currentUser = userList.Find(user => user.username == username && user.password == Crypto.CalculateMD5Hash(password));
-            return currentUser;
+
+            if (userList.Exists(x => x.username == username))
+            {
+                return false;
+            }
+            else
+            {
+                User user = new User(username, Crypto.CalculateMD5Hash(password));
+                userList.Add(user);
+                return Serializer.Save(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "/users.bin", userList);
+            }
         }
     }
 }
